design: 최종이용요금 computed 삭제 및 PayPriceDTO chargedEnergy type long으로 변경

+++ ifive/src/main/java/kr/co/milionvolt/ifive/domain/payment/PayPriceDTO.java
@@ -17,7 +17,7 @@ public class PayPriceDTO {
     private int stationId;
     private int chargerId;
     private long amount;
-    private long chargedEnergy;
+    private double chargedEnergy;
     private String paymentMethod;
     private String paymentStatus;
     private LocalDateTime createdAt;
